reorganized some of the utils in their own directory
