+++ src/lib.rs
@@ -1,11 +1,7 @@
 pub mod alert;
 pub mod conf;
-pub mod db;
-mod fake_ml_worker;
 pub mod filter;
-pub mod fits;
 pub mod kafka;
+pub mod ml;
 pub mod scheduling;
-pub mod spatial;
-pub mod testing_util;
-pub mod worker_util;
+pub mod utils;
