+++ src/utils/spatial.rs
