+++ src/utils/worker.rs
@@ -1,6 +1,4 @@
 use config::Config;
-use redis::streams::{StreamReadOptions, StreamReadReply};
-use redis::AsyncCommands;
 use std::{
     fmt,
     sync::{Arc, Mutex},
@@ -41,42 +39,6 @@ pub fn check_flag(flag: Arc<Mutex<bool>>) -> bool {
         }
         _ => false,
     }
-}
-
-pub async fn get_candids_from_stream(
-    con: &mut redis::aio::MultiplexedConnection,
-    stream: &str,
-    options: &StreamReadOptions,
-) -> Vec<i64> {
-    let result: Option<StreamReadReply> = con
-        .xread_options(&[stream.to_owned()], &[">"], options)
-        .await
-        .unwrap();
-    let mut candids: Vec<i64> = Vec::new();
-    if let Some(reply) = result {
-        for stream_key in reply.keys {
-            let xread_ids = stream_key.ids;
-            for stream_id in xread_ids {
-                let candid = stream_id.map.get("candid").unwrap();
-                // candid is a Value type, so we need to convert it to i64
-                match candid {
-                    redis::Value::BulkString(x) => {
-                        // then x is a Vec<u8> type, so we need to convert it an i64
-                        let x = String::from_utf8(x.to_vec())
-                            .unwrap()
-                            .parse::<i64>()
-                            .unwrap();
-                        // append to candids
-                        candids.push(x);
-                    }
-                    _ => {
-                        warn!("Candid unknown type: {:?}", candid);
-                    }
-                }
-            }
-        }
-    }
-    candids
 }
 
 pub fn get_check_command_interval(conf: Config, stream_name: &str) -> i64 {
